add request api

+++ database/migrations/2025_01_26_003859_create_requests_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->text('body')->nullable();
 
             $table->boolean('confirm')->default(false);
-            $table->integer('amount')->nullable();
+            $table->unsignedBigInteger('amount')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
